requirements.txt
Flask==3.0.2
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flask
psycopg2-binary
python-dotenv
requests
vercel-blob